Views now use blueprints. Removed circular imports.
