+++ app/views/__init__.py
@@ -1,0 +1,12 @@
+# from app import app, login_manager, bcrypt, db, models
+# from flask import render_template, flash, redirect, session, url_for, request, g, abort, Blueprint
+# from ..forms import *
+# from flask_login import login_user, logout_user, current_user, login_required
+# from werkzeug.utils import secure_filename
+# from datetime import datetime
+# import os
+
+
+
+
+
