+++ run.py
@@ -1,3 +1,3 @@
-#!flask/bin/python
+#!flask/Scripts/python
 from app import app
 app.run(debug=True)
